# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIBE Social Discovery App API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  age: number;
  bio: string;
  photoUrl: string;
  isPremium: boolean;
  city?: string;
  interests?: string[];
}

export interface MatchedUser {
  id: number;
  name: string;
  age: number;
  bio: string;
  photoUrl: string;
  isPremium: boolean;
  city?: string;
  matchId: number;
  lastMessage?: string;
  unreadCount: number;
}

export interface RegisterRequest {
  name: string;
  age: number;
  bio: string;
  photoUrl?: string;
  city?: string;
  interests?: string[];
}

export interface UpdateUserRequest {
  name?: string;
  age?: number;
  bio?: string;
  photoUrl?: string;
  city?: string;
  interests?: string[];
}

export type CreateLikeRequestAction =
  (typeof CreateLikeRequestAction)[keyof typeof CreateLikeRequestAction];

export const CreateLikeRequestAction = {
  like: "like",
  dislike: "dislike",
} as const;

export interface CreateLikeRequest {
  fromUserId: number;
  toUserId: number;
  action: CreateLikeRequestAction;
}

export interface LikeResult {
  success: boolean;
  isMatch: boolean;
}

export interface FriendRequest {
  fromUserId: number;
  toUserId: number;
}

export interface SendMessageRequest {
  senderId: number;
  receiverId: number;
  content: string;
}

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  content: string;
  isRead: boolean;
  createdAt?: string;
}

export interface UnreadCount {
  count: number;
}

export interface ActivatePremiumRequest {
  userId: number;
}

export type GetUsersParams = {
  currentUserId: number;
};