# lib/api-zod/src/generated/types/createLikeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIBE Social Discovery App API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateLikeRequestAction } from "./createLikeRequestAction";

export interface CreateLikeRequest {
  fromUserId: number;
  toUserId: number;
  action: CreateLikeRequestAction;
}