# lib/api-zod/src/generated/types/likeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIBE Social Discovery App API
 * OpenAPI spec version: 0.2.0
 */

export interface LikeResult {
  success: boolean;
  isMatch: boolean;
}