# lib/api-zod/src/generated/types/matchedUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIBE Social Discovery App API
 * OpenAPI spec version: 0.2.0
 */

export interface MatchedUser {
  id: number;
  name: string;
  age: number;
  bio: string;
  photoUrl: string;
  isPremium: boolean;
  city?: string;
  matchId: number;
  lastMessage?: string;
  unreadCount: number;
}